backend/pyproject.toml
[project]
name = "crucible-eval-backend"
version = "0.1.0"
description = "FastAPI backend for Crucible Eval"
requires-python = ">=3.12"
dependencies = [
  "anthropic==0.34.2",
  "fastapi==0.115.0",
  "google-genai==0.3.0",
  "httpx>=0.27,<0.28",
  "openai==1.51.0",
  "pydantic==2.8.2",
  "python-dotenv==1.0.1",
  "PyYAML==6.0.2",
  "uvicorn[standard]==0.30.6",
]

[tool.uv]
package = false
